dispatcher/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
kubernetes>=28.0.0
pyyaml>=6.0
